pp_dbstate: skip SAVESTACK_POS()
On PERL_RC_STACK builds, skip the call to SAVESTACK_POS().
I suspect that saving the stack position is no longer
required. It was added in 5.001 by:

    NETaa13155: &DB::DB left trash on the stack.
    From: Thomas Koenig
    Files patched: lib/perl5db.pl pp_ctl.c
     The call by pp_dbstate() to &DB::DB left trash on the
     stack.  It now calls DB in list context, and DB returns
     ().

but the details of what bug it fixed are long lost to history.
SAVESTACK_POS() doesn't work well with stacks which may be split into
partly reference-counted and partly not halves, so skip it and hope it
doesn't cause any problems.

Also, assert that SAVEt_STACK_POS isn't used anywhere any more.