Release schedule for 5.39 dev cycle
Skeleton only; no assignments yet.

NOTE: The 5.38.0 release manager opted not to do a "same day" or "next
day" of perl-5.39.0.  Does that mean that we start this year's dev cycle
with a 5.39.0 release on July 20 --or a 5.39.1?  (Assuming the latter.)

NOTE: As of the time of creation of this branch, having no 5.39.0
release we also lack a v5.39.0 annotated tag.

This has implications for CPANtesters reports.  Task::CPAN::Reporter,
for example, by default does not transmit multiple PASS reports for a
given version of a given distro for a given Perl release.  If I tested
WWW-Mechanize-2.17 against one of the perl-5.38.0-RC* releases, I can't
transmit a new report until we're at 5.39.*.